# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Strait API
  version: 1.0.0
  description: A production-grade Go job strait API.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /v1/api-keys:
    get:
      tags:
        - APIKeys
      summary: List API keys
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKey'
    post:
      tags:
        - APIKeys
      summary: Create an API key
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
  /v1/api-keys/{keyID}:
    delete:
      tags:
        - APIKeys
      summary: Revoke an API key
      security:
        - bearerAuth: []
      parameters:
        - name: keyID
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: API key revoked
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    runToken:
      type: http
      scheme: bearer
      description: JWT run token issued by the trigger response.
  schemas:
...
    APIKey:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        name:
          type: string
        key_prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
...